rename: aepctl → aectl across the entire codebase
Renames the CLI tool directory, Go module path, all internal imports,
Kubernetes resource strings (field manager, annotation key, labels),
Helm chart references, GitHub Actions workflow paths, and docs.
