+++ deployments/helm-charts/platform/templates/thunder-app-operator/thunder-applications.yaml
@@ -1,7 +1,7 @@
 # Platform Thunder OAuth clients provisioned by the thunder-app-operator.
 #
 # Confidential clients (client_credentials): secrets are pre-generated by
-# `aepctl platform install` and synced from OpenBao via ESO into the
+# `aectl platform install` and synced from OpenBao via ESO into the
 # aep-thunder-secrets Secret in this namespace. The operator reads the secret
 # when reconciling each CR — if the Secret is not yet synced, the CR retries
 # with a 30s backoff until ESO delivers it.
