+++ deployments/helm-charts/platform/values.yaml
@@ -180,7 +180,7 @@ thunder:
   jwksURL: "http://thunder-service.thunder.svc.cluster.local:8090/oauth2/jwks"
 
 # thunder-app-operator subchart values.
-# enabled=true (default): operator runs as a subchart — used by aepctl platform install.
+# enabled=true (default): operator runs as a subchart — used by aectl platform install.
 # enabled=false: subchart is skipped — used by local dev (setup-local.sh installs the
 # operator as a standalone Helm release in thunder-app-operator-system instead).
 # ThunderApplication CRs in templates/thunder-app-operator/ are always deployed
@@ -189,7 +189,7 @@ thunder-app-operator:
   enabled: true
   thunder:
     adminURL: "http://thunder-service.thunder.svc.cluster.local:8090"
-    # existingSecret is set by aepctl platform install via --set to point at the
+    # existingSecret is set by aectl platform install via --set to point at the
     # ESO-synced aep-thunder-operator-creds Secret. Local dev leaves this empty and
     # uses the chart-managed Secret with the operator chart's built-in defaults.
