+++ docs/superpowers/specs/2026-07-21-skills-experience-design.md
@@ -135,7 +135,7 @@ The shared artifact is a **contract** — the manifest schema, the content-hash 
 | Org reconcile (platform → org repo, 3-way) | BFF, server-side | Go — existing `Reconcile()` behind `POST /org/skills/sync` | settings-page button, project creation, platform release | core |
 | Project seed/refresh (org repo → project skill copies) | BFF, server-side — the **single writer** of skill copies | Go — same committed-truth write surface as design files; diff-first, commits to main | project creation; design save/publish; **task dispatch** (guarantees fresh copies in every task clone) | core |
 | Runner consumption (no sync) | runner pod | TS — reads `.claude/skills/` from its project clone; preload pinned via SDK `skills:` array; on-demand for the rest; records loads | every task — zero skill-related network; the org-skills clone of today's `skills_resolver.ts` is retired | core |
-| Local freshness (org repo → dev clone) | developer machine | Go — new `aep skills sync` command in the existing `tools/aepctl` CLI (binary `aep`, already has `login` + `~/.aep/config.yaml`); writes the project skill copies + `CLAUDE.md` "load these first" note | manual, **optional** | convenience — ship last |
+| Local freshness (org repo → dev clone) | developer machine | Go — new `aep skills sync` command in the existing `tools/aectl` CLI (binary `aep`, already has `login` + `~/.aep/config.yaml`); writes the project skill copies + `CLAUDE.md` "load these first" note | manual, **optional** | convenience — ship last |
 
 The CLI is never packed into project repos and is not required for local dev: a fresh clone already carries the full library from the last refresh commit — `aep skills sync` only closes the staleness window between refresh commits. Project repos carry only data: `skillsApplied[]` (already committed) + the skill copies.
