+++ tools/aectl/cmd/platform.go
