+++ tools/aectl/cmd/secret_import.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/wso2/aep/aepctl/internal/openbao"
+	"github.com/wso2/aep/aectl/internal/openbao"
 )
 
 var (
