+++ tools/aectl/cmd/sre_assets.go
