+++ tools/aectl/internal/thunder/authscript.go
