+++ tools/aectl/internal/thunder/cors.go
